chore: refactor old code + use new lenis + update sitemap + update to use pnpm
