+++ .gitignore
@@ -25,3 +25,4 @@ yarn-error.log*
 
 
 .vscode/
+.vercel
