+++ package.json
@@ -7,13 +7,9 @@
     "@babel/core": ">=7.0.0-0 <8.0.0",
     "@babel/plugin-syntax-flow": "7.14.5",
     "@babel/plugin-transform-react-jsx": "7.14.9",
-    "@studio-freight/lenis": "^0.2.28",
-    "@testing-library/dom": "^8.20.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.2.1",
     "firebase": "^9.8.3",
     "gsap": "^3.11.4",
+    "lenis": "^1.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "5.0.0",
@@ -28,9 +24,15 @@
   },
   "eslintConfig": {
     "extends": [
-      "react-app",
-      "react-app/jest"
+      "react-app"
     ]
+  },
+  "devDependencies": {
+    "autoprefixer": "10.4.2",
+    "postcss": "8.4.6",
+    "prettier": "2.5.1",
+    "prettier-plugin-tailwindcss": "0.1.5",
+    "tailwindcss": "3.0.18"
   },
   "browserslist": {
     "production": [
@@ -43,15 +45,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "autoprefixer": "10.4.2",
-    "postcss": "8.4.6",
-    "prettier": "2.5.1",
-    "prettier-plugin-tailwindcss": "0.1.5",
-    "tailwindcss": "3.0.18"
-  },
-  "engines": {
-    "node": ">=12.0.0"
   }
 }
