+++ public/sitemap.xml
@@ -8,8 +8,8 @@
 
 
 <url>
-  <loc>https://cinematrix.subzeroo.tech/</loc>
-  <lastmod>2022-02-25T06:35:28+00:00</lastmod>
+  <loc>https://cinematrix-three.vercel.app/</loc>
+  <lastmod>2024-06-18T16:42:08+00:00</lastmod>
 </url>
