+++ src/index.jsx
@@ -2,11 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import "./index.css";
-import App from "./App";
+import BaseApp from "./BaseApp";
 
-import Lenis from '@studio-freight/lenis'
-
-// GOogle Analytics
+// Google Analytics
 import { initializeApp } from "firebase/app";
 import { getAnalytics } from "firebase/analytics";
 import { getPerformance } from "firebase/performance";
@@ -27,31 +25,12 @@ const app = initializeApp(firebaseConfig);
 getAnalytics(app);
 getPerformance(app);
 
-const lenis = new Lenis({
-	duration: 1.5,
-	easing: (t) => Math.min(1, 1.001 - Math.pow(2, -10 * t)), // https://www.desmos.com/calculator/brs54l4xou
-	direction: "vertical", // vertical, horizontal
-	gestureDirection: "vertical", // vertical, horizontal, both
-	smooth: true,
-	mouseMultiplier: 1.2,
-	smoothTouch: true,
-	touchMultiplier: 2,
-	infinite: false,
-});
 
-
-
-function raf(time) {
-	lenis.raf(time);
-	requestAnimationFrame(raf);
-}
-
-requestAnimationFrame(raf);
 
 
 ReactDOM.render(
     <React.StrictMode>
-        <App />
+        <BaseApp />
     </React.StrictMode>,
     document.getElementById("root")
 );
