+++ src/methods/fetchData.js
@@ -1,0 +1,8 @@
+async function fetchData(input) {
+    const data = await fetch(process.env.REACT_APP_URL + '?movie=' + input);
+
+    if (!data.ok) return [];
+    return data.json();
+}
+
+export default fetchData;
